get chemical environments from a full search in fingerprint
I thought it was okay to use the more basic/faster search instead of a full
labeling here because the initial ChEMBL search had already determined that
these match. However, I forgot that my splitting on '.' and now fragmenting
could alter the labeling results.

This was leading to problems when trying to check my results in Python as part
of generating a final data set.